requirements.txt
openai==0.27.8
pandas==2.0.2
Pillow==9.5.0
Pillow==9.5.0
streamlit==1.23.1
streamlit_chat==0.0.2.2
